* disabling direct_io mode in case of MAC OS X by default.
* removed warnings in unify,posix,client-protocol.

* added header files in glusterfs-guts/Makefile.am

+++ xlators/cluster/unify/src/unify.c
@@ -3596,7 +3596,7 @@ unify_rename (call_frame_t *frame,
     for (index = 0; index <= priv->child_count ; index++) {
       _STACK_WIND (frame, 
 		   unify_rename_lookup_cbk,
-		   (void *)index, 
+		   (void *)(long)index, 
 		   priv->xl_array[index],
 		   priv->xl_array[index]->fops->lookup,
 		   &tmp_loc,
